feat(soql): move mehtod to common utils
