+++ packages/salesforcedx-vscode-soql/src/commonUtils.ts
